build: Fix `make term` creating new child terminal windows without side effects
Previously, commit 20d2e60 caused changes header files to not trigger a build of related sourcefiles.

The change to the dependency files inclusion is needed to still trigger a build of related source files when a header file changes.
Technically, this is only needed when `MAKECMDGOALS` would not be exported anymore, but it's still nice to have when doing `make <program_name>`.